+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## Added
+
+- Azurerm now can use the `--tags` filter
+  ([Issue #361](https://github.com/cycloidio/terracognita/issues/361))
+
 ## [0.8.3] _2023-03-14_
 
 ### Fixed